python_backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
yagmail
jinja2
